+++ src/main/java/in/bhargavrao/stackoverflow/natty/roomdata/RPublicChatRoom.java
@@ -4,12 +4,8 @@ import fr.tunaki.stackoverflow.chat.Room;
 import fr.tunaki.stackoverflow.chat.event.MessageReplyEvent;
 import fr.tunaki.stackoverflow.chat.event.UserMentionedEvent;
 import in.bhargavrao.stackoverflow.natty.commandlists.RPublicCommandsList;
-import in.bhargavrao.stackoverflow.natty.commandlists.SoBoticsCommandsList;
 import in.bhargavrao.stackoverflow.natty.printers.PostPrinter;
 import in.bhargavrao.stackoverflow.natty.printers.RPublicPostPrinter;
-import in.bhargavrao.stackoverflow.natty.printers.SoBoticsPostPrinter;
-import in.bhargavrao.stackoverflow.natty.utils.PostUtils;
-import in.bhargavrao.stackoverflow.natty.validators.AllowAllNewAnswersValidator;
 import in.bhargavrao.stackoverflow.natty.validators.RPublicValidator;
 import in.bhargavrao.stackoverflow.natty.validators.Validator;
 
@@ -20,7 +16,7 @@ import java.util.function.Consumer;
  */
 public class RPublicChatRoom implements BotRoom{
     @Override
-    public long getRoomId() {
+    public int getRoomId() {
         return 25312;
     }
